+++ resources/js/components/PreviewScreen.vue
@@ -63,13 +63,6 @@
             command() {
                 return _.find(this.batch, {type: 'command'})
             },
-
-            gravatarUrl() {
-                if (this.entry.content.user.email) {
-                    const md5 = require('md5')
-                    return 'https://www.gravatar.com/avatar/' + md5(this.entry.content.user.email.toLowerCase()) + '?s=200'
-                }
-            }
         },
 
 
@@ -242,7 +235,7 @@
                     <td class="table-fit font-weight-bold align-middle">Name</td>
 
                     <td class="align-middle">
-                        <img :src="gravatarUrl" class="mr-2 rounded-circle" height="40" width="40" v-if="gravatarUrl">
+                        <img :src="entry.content.user.avatar" :alt="entry.content.user.name" class="mr-2 rounded-circle" height="40" width="40" v-if="entry.content.user.avatar">
                         {{entry.content.user.name}}
                     </td>
                 </tr>