+++ resources/js/screens/requests/preview.vue
@@ -58,6 +58,13 @@
         </tr>
 
         <tr>
+            <td class="table-fit font-weight-bold">IP Address</td>
+            <td>
+                {{slotProps.entry.content.ip_address || '-'}}
+            </td>
+        </tr>
+
+        <tr>
             <td class="table-fit font-weight-bold">Memory usage</td>
             <td>
                 {{slotProps.entry.content.memory || '-'}} MB